Split llg_test_utils/mod.rs into trace_check and acceptance submodules
Split the 648-line mod.rs into three files by testing strategy:
- mod.rs: shared PARSER_FACTORY state and pub use re-exports
- trace_check.rs: trace-replay grammar checking engine
- acceptance.rs: feed-and-check string/schema validation

Pure code-motion refactor; no behavioral changes.
